refactor: Router 컴포넌트 파일명 컨벤션에 맞게 변경
