+++ src/router/index.ts
@@ -1,0 +1,3 @@
+import Router from './Router';
+
+export default Router;
